+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * fix: WEC failed to extract links from some websites, error: Cannot read properties of undefined (reading 'startsWith')
 * dependencies: overwrite the dependency of `puppeteer-har` to address `npm audit` warnings
+* feature: support collection of whitelisted hosts in the HTTP header Content-Security-Policy (CSP)
 
 ## 2.1.2 / 2024-04-25